Library Functions

FMIFlux.CS_NeuralFMU

FMIFlux.ME_NeuralFMU

FMIFlux.NeuralFMU

FMIFlux.fmi2EvaluateME

FMIFlux.fmi2InputDoStepCSOutput

FMIFlux.fmiDoStepCS

FMIFlux.fmiEvaluateME

FMIFlux.fmiInputDoStepCSOutput

FMIFlux.mse_interpolate

FMIFlux functions

FMIFlux.CS_NeuralFMU — Type

Structure definition for a NeuralFMU, that runs in mode Co-Simulation (CS).

FMIFlux.ME_NeuralFMU — Type

Structure definition for a NeuralFMU, that runs in mode Model Exchange (ME).

FMIFlux.NeuralFMU — Type

The mutable struct representing an abstract (simulation mode unknown) NeuralFMU.

FMI 2 version dependent functions

Missing docstring.

Missing docstring for fmi2DoStepCS. Check Documenter's build log for details.

FMIFlux.fmi2EvaluateME — Function

DEPRECATED:

Performs something similar to fmiDoStep for ME-FMUs (note, that fmiDoStep is for CS-FMUs only). Event handling (state- and time-events) is supported. If you don't want events to be handled, you can disable event-handling for the NeuralFMU nfmu with the attribute eventHandling = false.

Optional, additional FMU-values can be set via keyword arguments setValueReferences and setValues. Optional, additional FMU-values can be retrieved by keyword argument getValueReferences.

Function takes the current system state array ("x") and returns an array with state derivatives ("x dot") and optionally the FMU-values for getValueReferences. Setting the FMU time via argument t is optional, if not set, the current time of the ODE solver around the NeuralFMU is used.

FMIFlux.fmi2InputDoStepCSOutput — Function

DEPRECATED:

fmi2InputDoStepCSOutput(comp::FMU2Component, 
                        dt::Real, 
                        u::Array{<:Real})

Sets all FMU inputs to u, performs a ´´´fmi2DoStep´´´ and returns all FMU outputs.

FMI version independent functions

FMIFlux.fmiDoStepCS — Function

DEPRECATED:

Wrapper. Call fmi2DoStepCS for more information.

FMIFlux.fmiEvaluateME — Function

DEPRECATED:

Wrapper. Call fmi2EvaluateME for more information.

FMIFlux.fmiInputDoStepCSOutput — Function

DEPRECATED:

Wrapper. Call fmi2InputDoStepCSOutput for more information.

Additional functions

FMIFlux.mse_interpolate — Function

Compares non-equidistant (or equdistant) datapoints by linear interpolating and comparing at given interpolation points t_comp. (Zygote-friendly: Zygote can differentiate through via AD.)

Missing docstring.

Missing docstring for transferParams!. Check Documenter's build log for details.